requirements.txt
# 基础依赖
streamlit>=1.30.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langgraph>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0

# 文档处理
pypdf>=4.0.0
markdown>=3.5.1
beautifulsoup4>=4.12.2
html2text>=2020.1.16

# 向量存储
faiss-cpu>=1.7.4

# 模型API客户端
openai>=1.6.0
zhipuai>=2.0.0
dashscope>=1.13.0
requests>=2.31.0

# 网页自动化
playwright>=1.40.0

# 其他工具
tqdm>=4.66.1